init-template: Explain how to bootstrap a debian template image
We will build debian vms (oldstable/stable) from a template with some basis
already set-up to avoid repeating those actions for new provisioned vm.

This action is needed to be once in a while (every debian migration to new
stable for example). Still, this is worth documenting/scripting to ease
reproducibility.

Related T1785