adapt the debian buster template to be used by terraform
- The current template is configured with a default lsi scsi adaptor
and a scsi disk. Terraform is using a virtio disk and creates a new
virtio disk one if not existing instead of extended the current
scsi volume. Configuring the template wirh a virtio disk solves
the problem.
- a qemu-server upgrade has changed the boot order option behavior which
led to an wrong definition of the boot devices by packer.
It forces to perform a manual action during the template creation
and to explicitly declare the cloud_init_storage_pool option
because the default value is computed from the now not existing boot
device. It will be fixed on a future packer release[1].

[1]: https://github.com/hashicorp/packer/pull/10208

Related to T2747