Increase size of the backup01 data disk
The attach create_option of the disk changed from "empty"
to "attach" after the manual actions (unattach/reattach it).
The only way I found to avoid an error is to ignore this field.
Unfortunately, terraform doesn't not support ignoring a dynamic
resource's field, so it has to be pointed directly with the index

Related to T4222