staging: Add the test cluster definition
This cluster was used to test the kubernetes upgrade before doing
it for real in the archive-staging-cluster.

Such cluster can be useful to test some workload or migrations
more easily than in a local cluster. So let's keep it for now.

The configuration allow the full creation in one plan/apply command
without human actions. It's possible with the definition of the
vmid parameter for the vms and the creation of the snapshotter configuration
file in the post provision step

Related to swh/infra/sysadm-environment#5014