k8s: Speedup the deployments
It's a workaround for the "failed to reserve container name" issue
during big deployments. Most of the pods fails to start and takes a lot
of time to recover.

The default value is 2mn which is too small when a big deployment is in
progress.

A lot of information about this issue is available at:
https://github.com/containerd/containerd/issues/4604