archive-production-rke2: Increase prometheus limits
Since a couple of days, the prometheus pods was regularly killed
and most of the time didn't succeed to start due to a startup probe
timeout too slow.

This commit address the root cause of the first kill.

A second one (later) will try to fix the too short startup probe delay which
avoid the pod to recover (database check too long after the crash)